Move ZMQ and SQLite into dpipes subpackages
