+++ dpipes/go.mod
@@ -2,10 +2,4 @@ module github.com/tmbdev/tarp/dpipes
 
 go 1.14
 
-require (
-	github.com/Masterminds/squirrel v1.2.0
-	github.com/mattn/go-sqlite3 v2.0.3+incompatible
-	github.com/shamaton/msgpack v1.1.1
-	github.com/stretchr/testify v1.2.2
-	gopkg.in/zeromq/goczmq.v4 v4.1.0
-)
+require github.com/stretchr/testify v1.5.1
